+++ src/consts.js
@@ -1,0 +1,19 @@
+export const ColorList = [
+  "hotpink",
+  "deepPink",
+  "MediumVioletRed",
+  "PaleVioletRed",
+  "Coral",
+  "Tomato",
+  "Orange",
+  "Chocolate",
+  "Peru",
+  "Sienna",
+  "SaddleBrown",
+  "Maroon",
+  "Crimson",
+];
+
+export const EventKey = {
+  ENTER: "Enter",
+};
